- added some mappers

+++ app/src/main/java/com/smart/resources/schools_app/features/onlineExam/data/remote/model/NetworkOnlineExam.kt
@@ -4,11 +4,11 @@ import org.threeten.bp.LocalDate
 import org.threeten.bp.LocalTime
 
 data class NetworkOnlineExam(
-    val idOnlineExam: Int?,
-    val teacherId: Int?,
+    val idOnlineExam: Long?,
+    val teacherId: Long?,
     val date: LocalDate?,
     val time: LocalTime?,
-    val examTime: Int?,
+    val examTime: Long?,
     val nuOfRequiredQuestion: Int?,
     val examKey: String?,
     val isFinished: Int?,
